fix(rmm-eac2): fix MISRA C:2012 Rule 10.3 in vmid.c

+++ runtime/core/vmid.c
@@ -41,8 +41,8 @@ bool vmid_reserve(unsigned int vmid)
 		return false;
 	}
 
-	offset = vmid / BITS_PER_UL;
-	vmid %= BITS_PER_UL;
+	offset = vmid / (unsigned int)BITS_PER_UL;
+	vmid %= (unsigned int)BITS_PER_UL;
 
 	return !atomic_bit_set_acquire_release_64(&vmids[offset], vmid);
 }
@@ -60,8 +60,9 @@ void vmid_free(unsigned int vmid)
 
 	/* Check the number of supported VMID values */
 	assert(vmid < vmid_count);
-	offset = vmid / BITS_PER_UL;
-	vmid %= BITS_PER_UL;
+
+	offset = vmid / (unsigned int)BITS_PER_UL;
+	vmid %= (unsigned int)BITS_PER_UL;
 
 	atomic_bit_clear_release_64(&vmids[offset], vmid);
 }
